requirements.txt
# Core dependencies
yfinance>=0.2.18
requests>=2.31.0
beautifulsoup4>=4.12.2
pandas>=2.0.3
feedparser>=6.0.10
python-dateutil>=2.8.2

# Database
pymongo>=4.5.0

# Machine Learning & Statistics
statsmodels>=0.14.0
scikit-learn>=1.3.0
tensorflow>=2.13.0
numpy>=1.24.3

# Visualization
plotly>=5.15.0

# Web Framework
flask>=2.3.2

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0

# Model Hub
huggingface_hub>=0.23.0